% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boxcox_transform}
\alias{boxcox_transform}
\title{Helper function to calculate optimal Box-Cox power transformation}
\usage{
boxcox_transform(x, plot = FALSE, family = c("bcPower", "yjPower"), ...)
}
\arguments{
\item{x}{numeric data vector}

\item{plot}{logical. Plot the BC's objective function? Default: no.}

\item{family}{transformation family to use: Box-Cox or Yeo-Johnson. The
options match those in \code{\link[car]{boxCox}}}

\item{...}{additional arguments passed to \code{\link[car]{boxCox}}}
}
\value{
list with the transformed data in \code{x} and the optimal BC value
in \code{lambda}.
}
\description{
A Box-Cox (BC) transformation is often used to transform data to achieve a
symmetric, unimodal distribution (typically, of the response). The BC
transformation can be estimated via \code{\link[car]{boxCox}}. The code to
find the optimal value is a bit cumbersome. In this function, we streamline
the process.
}
\details{
The BC transformation requires a positive response. In this case, the
Yeo-Johnson (YJ) transformation generalizes the BC transformation to allow
all real values.
}

