% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cut-pretty.r
\name{cut_pretty}
\alias{cut_pretty}
\title{Cuts a vector into factors with pretty levels}
\usage{
cut_pretty(x, breaks, collapse = " to ", ...)
}
\arguments{
\item{x}{numeric vectory}

\item{breaks}{numeric vector of two ore more unique cut points}

\item{...}{arguments passed to \code{\link[base]{cut}}}
}
\value{
A \code{\link{factor}} is returned
}
\description{
Cuts a vector into factors with pretty levels
}
\examples{
set.seed(42)
x <- runif(n=50, 0, 50)
cut_pretty(x, breaks=pretty(x))
}

