% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inv_boxcox}
\alias{inv_boxcox}
\title{Inverse Box-Cox Transformation}
\usage{
inv_boxcox(y, lambda = 0)
}
\arguments{
\item{y}{numeric vector}

\item{lambda}{Box-Cox transformation parameter}
}
\value{
transformed numeric vector
}
\description{
Often, a data transformation is necessary to achieve a symmetric, unimodal
distribution. A common approach is the Box-Cox transformation, which can be
performed by \code{\link[car]{boxCox}}. While the Box-Cox transformation is
given in the \code{\link{car}} package, the inverse transformation is not.
This function takes care of that.
}

