% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{pocketknife}
\alias{package-pocketknife}
\alias{pocketknife}
\alias{pocketknife-package}
\title{A collection of useful utility functions}
\description{
A collection of useful utility functions
}
\section{Package Description}{


The R package \code{pocketknife} contains a variety of useful utility
functions to simplify and to enhance common data-manipulation tasks in R.
}

