% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/retry.r
\name{retry}
\alias{retry}
\title{Retry expression with exponential backoff}
\usage{
retry(expr, silent = FALSE, max_attempts = 10, verbose = FALSE)
}
\arguments{
\item{expr}{an R expression to try.}

\item{silent}{logical: should the report of error messages be suppressed?}

\item{verbose}{logical: Should detailed messages be reported regarding each
attempt? Default: no.}

\item{max_tries}{the maximum number of times to attempt the expression
\code{expr}}
}
\value{
the value of the expression in \code{expr}. If the final attempt was
a failure, the objected returned will be of class try-error".
}
\description{
Attempts the expression in \code{expr} up to the number of tries specified in
\code{max_attempts}. Each time a failure results, the functions sleeps for a
random amount of time before re-attempting the expression. The upper bound of
the backoff increases exponentially after each failure.
}
\details{
For details on exponential backoff, see:
\url{http://en.wikipedia.org/wiki/Exponential_backoff}
}
\examples{
# Silly example that will never succeed.
retry(log("a"), verbose=TRUE, max_attempts=5)
}

