% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/table.r
\name{table_na}
\alias{table_na}
\title{Contingency table including NA values}
\usage{
table_na(...)
}
\arguments{
\item{...}{one or more objects passed to \code{\link[base]{table}} to build
a contingency table}
}
\value{
a \code{table} object containing the contingency table
}
\description{
This function is a simple wrapper around \code{\link[base]{table}} to build
a contingency table including \code{NA} values.
}

