% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pocketknife.r
\docType{package}
\name{pocketknife}
\alias{package-pocketknife}
\alias{pocketknife}
\alias{pocketknife-package}
\title{A collection of useful utility functions}
\description{
A collection of useful utility functions
}
\section{Package Description}{


The R package \code{pocketknife} contains a variety of useful utility
functions to simplify and to enhance common data-manipulation tasks in R.
}

